finance_agent/requirements.txt
google-auth[requests]>=2.47.0,<3.0.0
requests==2.32.4
